+++ frontend/src/pages/dashboard/jios/index.tsx
@@ -13,9 +13,11 @@ import JiosForm, { JioFormValues } from './JiosForm';
 import { useDispatch, useSelector } from '../../../store';
 import { setJioFormValues } from '../../../store/reducers/jioFormValues';
 import { setDateTime } from '../../../utils/formatTime';
-import { createJio, updateJio } from 'src/services/jios';
 import { useRequest } from 'ahooks';
 import { useSnackbar } from 'notistack';
+import JiosCreate from './JiosCreate';
+import JiosEdit from './JiosEdit';
+import JiosSearch from './JiosSearch';
 
 const StyledTab = styled(Tab)({
   '&.MuiButtonBase-root': {
@@ -31,36 +33,10 @@ enum TabValue {
 export default function Jios() {
   const TABS: TabValue[] = [TabValue.Open, TabValue.MyJios];
 
-  const dispatch = useDispatch();
   const refresh = useRefresh();
   const navigate = useNavigate();
   const [tabValue, setTabValue] = useState<TabValue>(TabValue.Open);
-  const { enqueueSnackbar } = useSnackbar();
   const jioFormValues = useSelector((state) => state.jioFormValues.data);
-
-  const { run: submitCreateJio } = useRequest(createJio, {
-    manual: true,
-    onSuccess: () => {
-      enqueueSnackbar('Created!');
-      refresh();
-      navigate('');
-    },
-    onError: (error) => {
-      // TODO: handle FE validation checks
-      enqueueSnackbar('Failed to create', { variant: 'error' });
-    },
-  });
-  const { run: submitUpdateJio } = useRequest(updateJio, {
-    manual: true,
-    onSuccess: () => {
-      enqueueSnackbar('Updated!');
-      refresh();
-      navigate('');
-    },
-    onError: () => {
-      enqueueSnackbar('Failed to update', { variant: 'error' });
-    },
-  });
 
   const handleRefresh = () => {
     refresh();
@@ -68,21 +44,6 @@ export default function Jios() {
 
   const onClickSearch = () => {
     navigate('search');
-  };
-
-  const handleSearch = async (data: JioFormValues) => {
-    dispatch(setJioFormValues(data));
-    navigate('');
-  };
-
-  const handleCreate = async (data: JioFormValues) => {
-    submitCreateJio(data);
-  };
-
-  const handleEdit = async (data: JioFormValues) => {
-    // FIXME: replace dummy id
-    submitUpdateJio(data, 0);
-    // navigate('');
   };
 
   const getStartDateTimeString = () => {
@@ -103,38 +64,9 @@ export default function Jios() {
 
   return (
     <Routes>
-      <Route
-        path="search"
-        element={
-          <JiosForm
-            isSearch
-            onSubmit={handleSearch}
-            submitLabel="Search"
-            submitIcon={<Iconify icon={'eva:search-outline'} width={24} height={24} />}
-            defaultValues={jioFormValues || undefined}
-          />
-        }
-      />
-      <Route
-        path="create"
-        element={
-          <JiosForm
-            onSubmit={handleCreate}
-            submitLabel="Submit"
-            submitIcon={<Iconify icon={'eva:add-outline'} width={24} height={24} />}
-          />
-        }
-      />
-      <Route
-        path="edit/:id"
-        element={
-          <JiosForm
-            onSubmit={handleEdit}
-            submitLabel="Submit"
-            submitIcon={<Iconify icon={'eva:add-outline'} width={24} height={24} />}
-          />
-        }
-      />
+      <Route path="search" element={<JiosSearch />} />
+      <Route path="create" element={<JiosCreate />} />
+      <Route path="edit/:id" element={<JiosEdit />} />
       <Route
         path=""
         element={
